refactor: move convertLineSpec to a legend utils file
